project_code/requirements.txt
# python 3.8
# conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch

pytorch-lightning==1.5.1
torchmetrics==0.6.0
opencv-python==4.4.0.44
wandb==0.12.7
pybullet==3.2.0
PyOpenGL-accelerate==3.1.5
scipy==1.7.3
torchvision==0.11.2
opencv-contrib-python

#pyexr==0.3.9
#h5py==3.6.0
#open3d==0.13.0
#pyrender==0.1.45
